moved profile route from app.py to profile.py
